I added search figure by point function and delete figures functions. Also was fixed some bugs.

+++ point.cpp
@@ -6,7 +6,7 @@ Point::Point()
     this->y = 0;
 }
 
-Point::Point(int x, int y)
+Point::Point(float x, float y)
 {
     this->x = x;
     this->y = y;
